feat: Add Ollama chat module and roadmap commands.
refactor(desktop): implement modular structure with shared dependencies

- Create new 'common' crate for shared utilities and configurations
- Centralize API_BASE_URL, CLIENT, and common error types in 'common' crate
- Update workspace Cargo.toml to include new 'common' crate
- Refactor existing crates to use shared utilities from 'common'
- Optimize dependency management across the project
- Improve code organization and reusability
