+++ Cargo.toml
@@ -1,7 +1,26 @@
 [workspace]
-members = ["apps/desktop/src-tauri"]
+members = [
+    "apps/desktop/src-tauri",
+    "apps/desktop/src-tauri/crates/common",
+    "apps/desktop/src-tauri/crates/auth",
+    "apps/desktop/src-tauri/crates/education",
+    "apps/desktop/src-tauri/crates/ai_chat",
+    "apps/desktop/src-tauri/crates/error",
+
+]
 resolver = "2"
 
+[workspace.dependencies]
+serde_json = "1.0"
+serde = { version = "1.0", features = ["derive"] }
+log = "0.4.22"
+tauri = { version = "2.0.3", features = ["protocol-asset"] }
+tauri-plugin-shell = "2.0.1"
+tauri-plugin-store = "2.0.1"
+reqwest = { version = "0.12.8", features = ["json"] }
+thiserror = "1.0.64"
+anyhow = "1.0.89"
+tokio = { version = "1.40.0", features = ["full"] }
 [profile.dev]
 incremental = true # Compile your binary in smaller steps.
