+++ apps/desktop/src-tauri/crates/ai_chat/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "ai_chat"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+log = { workspace = true }
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+ollama-rs = { version = "0.2.1", features = ["stream"] }
+scraper = "0.20.0"
+error = { path = "../error" }
+tauri-plugin-shell =  {workspace = true}
+tauri-plugin-store = {workspace = true}
+tauri = { workspace = true }
+chrono = { version = "0.4.38",features = ["serde"] }
+reqwest = {workspace = true}
+futures = "0.3.31"
+regex = "1.11.0"
