+++ apps/desktop/src-tauri/crates/auth/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod commands;
+
