+++ apps/desktop/src-tauri/crates/common/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "common"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { workspace = true }
+reqwest = { workspace = true }
