+++ apps/desktop/src-tauri/crates/common/src/state.rs
