+++ apps/desktop/src-tauri/crates/education/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "education"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+log = {workspace = true}
+serde = {workspace = true}
+tauri = {workspace = true}
+serde_json = {workspace = true}
+error = {path = "../error"}
+common = {path = "../common"}
