+++ apps/desktop/src-tauri/crates/education/src/courses/commands/mod.rs
