+++ apps/desktop/src-tauri/crates/education/src/exercises/commands/mod.rs
