+++ apps/desktop/src-tauri/crates/error/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "error"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow= {workspace = true}
+thiserror= {workspace = true}
+reqwest = {workspace = true}
+serde_json = {workspace = true}
+serde = {workspace = true}
+tauri = {workspace = true}
+tauri-plugin-store = {workspace = true}
+tauri-plugin-shell = {workspace = true}
