+++ apps/desktop/src-tauri/crates/error/src/lib.rs
