--- apps/desktop/src-tauri/src/ai/mod.rs
@@ -1,24 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-pub mod ollama;
-pub mod ollama_models;
-
-pub mod commands;
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct OllamaModelInfo {
-    name: String,
-    id: String,
-    size: String,
-    modified: String,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct OllamaModelDetails {
-    architecture: String,
-    parameters: String,
-    context_length: String,
-    embedding_length: String,
-    quantization: String,
-    license: String,
-}
