--- apps/desktop/src-tauri/src/education/mod.rs
@@ -1,6 +1,0 @@
-pub mod exercises;
-pub mod models;
-pub mod roadmaps;
-pub mod courses;
-pub mod lessons;
-pub mod modules;
